+++ test/unit/gemm/device/testbed_splitk.h
@@ -72,7 +72,7 @@ struct TestbedSplitK : public Testbed<Gemm> {
     // Determine SMEM requirements and waive if not satisfied
     //
 
-    int smem_size = int(sizeof(typename Gemm::GemmKernel::SharedStorage));
+    size_t smem_size = sizeof(typename Gemm::GemmKernel::SharedStorage);
 
     cudaDeviceProp properties;
     int device_idx;
